+++ app/src/main/java/com/example/myapplication/data/dao/UserDao.kt
@@ -14,7 +14,7 @@ interface UserDao {
     suspend fun getUserByEmail(email: String): User?
 
     @Query("SELECT * FROM user WHERE id = :userId")
-    suspend fun getUserById(userId: Int): User?
+    suspend fun getUserById(userId: Long): User?
 
     @Query("SELECT * FROM user WHERE name LIKE '%' || :query || '%'")
     fun searchByName(query: String): List<User>
