+++ app/src/main/java/com/example/myapplication/data/entities/Pet.kt
@@ -1,6 +1,5 @@
 package com.example.myapplication.data.entities
 
-import androidx.room.ColumnInfo
 import androidx.room.Entity
 import androidx.room.ForeignKey
 import androidx.room.PrimaryKey
@@ -18,8 +17,8 @@ import androidx.room.PrimaryKey
 )
 data class Pet(
     @PrimaryKey(autoGenerate = true)
-    val pet_id: Int = 0,
-    val userId: Int,
+    val pet_id: Long = 0,
+    val userId: Long,
     val name: String,
     val weight: Float,
     val google_drive_link: String? = null
